+++ zigar-compiler/zig/stub-napi.zig
@@ -1,3 +1,9 @@
+const std = @import("std");
+
+const module = @import("module");
+
 pub const host = @import("./host-napi.zig");
 
-export const zig_module = host.createModule(@import("module"));
+export const zig_module = host.createModule(module);
+
+pub const std_options: std.Options = if (@hasDecl(module, "std_options")) module.std_options else .{};